update int type

+++ python/ray/includes/libcoreworker.pxd
@@ -300,7 +300,7 @@ cdef extern from "ray/core_worker/core_worker.h" nogil:
         c_string serialized_job_config
         int metrics_agent_port
         c_bool connect_on_start
-        int runtime_env_hash
+        int64_t runtime_env_hash
         int worker_shim_pid
 
     cdef cppclass CCoreWorkerProcess "ray::CoreWorkerProcess":
